+++ docker/src/main/java/net/oneandone/stool/kubernetes/ContainerInfo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.oneandone.stool.docker;
+package net.oneandone.stool.kubernetes;
 
 public class ContainerInfo {
     public final String id;