+++ docker/src/main/java/net/oneandone/stool/kubernetes/FileNodes.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.oneandone.stool.docker;
+package net.oneandone.stool.kubernetes;
 
 import net.oneandone.sushi.fs.file.FileNode;
 import org.kamranzafar.jtar.TarEntry;