+++ server/src/main/java/net/oneandone/stool/server/configuration/ServerConfiguration.java
@@ -15,7 +15,7 @@
  */
 package net.oneandone.stool.server.configuration;
 
-import net.oneandone.stool.docker.Engine;
+import net.oneandone.stool.kubernetes.Engine;
 import net.oneandone.stool.server.util.Mailer;
 import net.oneandone.stool.server.util.Pool;