+++ server/src/main/java/net/oneandone/stool/server/util/Pool.java
@@ -15,8 +15,8 @@
  */
 package net.oneandone.stool.server.util;
 
-import net.oneandone.stool.docker.Engine;
-import net.oneandone.stool.docker.ServiceInfo;
+import net.oneandone.stool.kubernetes.Engine;
+import net.oneandone.stool.kubernetes.ServiceInfo;
 import net.oneandone.stool.server.stage.Stage;
 
 import java.io.IOException;